Keep buttonClass out of the component file
react-doctor: a file that exports both a component and a plain function
loses Fast Refresh, so editing either one full-reloads the page instead of
preserving state.
