+++ src/app/routes/domains.tsx
@@ -7,7 +7,8 @@ import { Trash2, RefreshCw, Star } from "lucide-react";
 import { useCurrentUser, useConfig, useDomains, useDomainMutations } from "../lib/hooks";
 import { useCurrentOrg } from "../lib/current-org";
 import { PLAN_LIMITS, type DomainDTO, type OrgRole } from "@/shared/types";
-import { Button, buttonClass, IconButton } from "../ui/button";
+import { Button, IconButton } from "../ui/button";
+import { buttonClass } from "../ui/button-class";
 import { Input } from "../ui/field";
 import { ConfirmDialog } from "../ui/confirm-dialog";
 import { withErrorToast } from "../lib/mutation-toast";
